feat: BBAnimatable 애니메이션 관련 Protocol  메서드 정의 및 내부 메서드 구현 - BBDrawable draw 관련 Protocol 메서드 정의 및 내부 메서드 구현 - BBToolTipConfig, BBToolTipPosition, BBToolTipType 파일 추가 - BBToolTipView BBDrawable, BBAnimatable 프로토콜 채택 코드 추가
