+++ 14th-team5-iOS/Core/Sources/Bibbi/BBCommons/BBToolTip/BBToolTipPosition.swift
@@ -1,0 +1,22 @@
+//
+//  BBToolTipPosition.swift
+//  Core
+//
+//  Created by Kim dohyun on 9/19/24.
+//
+
+import Foundation
+
+
+public enum BBToolTipXPosition: CGFloat {
+    case left = 0.0
+    case midLeft = 0.25
+    case center = 0.5
+    case midRight = 0.75
+    case right = 1.0
+}
+
+public enum BBToolTipPosition {
+    case top
+    case bottom
+}
